+++ config/config_test.go
@@ -100,6 +100,13 @@ func TestConfigValidate_varDefaultBadType(t *testing.T) {
 	}
 }
 
+func TestConfigValidate_varDefaultInterpolate(t *testing.T) {
+	c := testConfig(t, "validate-var-default-interpolate")
+	if err := c.Validate(); err == nil {
+		t.Fatal("should not be valid")
+	}
+}
+
 func TestProviderConfigName(t *testing.T) {
 	pcs := []*ProviderConfig{
 		&ProviderConfig{Name: "aw"},