+++ config/test-fixtures/validate-var-default-interpolate/main.tf
@@ -1,0 +1,3 @@
+variable "foo" {
+  default = "${aws_instance.foo.bar}"
+}